+++ tests/testthat/test-sortuse64.R
@@ -70,13 +70,11 @@ test_that("ordertab handles nunique smaller than actual", {
   x = c(x, x[1:8], x[1:6])
   o = order(x)
   # this should not segfault, even though nunique is wrong
-  gctorture(TRUE)
   for (ii in 1:10) {
     # with the fix, this will not crash, but return a truncated result
     # without the fix, this is likely to crash
-    res <- ordertab(x, o, 4)
+    res <- ordertab(x, o, 4L)
   }
-  gctorture(FALSE)
   # The result is truncated, but we can check it's not a crash and has the expected length
-  expect_length(res, 4)
+  expect_length(res, 4L)
 })